+++ guizero/CheckBox.py
@@ -1,6 +1,6 @@
 from tkinter import Checkbutton, IntVar
 from .mixins import WidgetMixin
-from .tkmixins import ScheduleMixin, DestroyMixin, EnableMixin, FocusMixin, DisplayMixin, ReprMixin
+from .tkmixins import ScheduleMixin, DestroyMixin, EnableMixin, FocusMixin, DisplayMixin, TextMixin, ReprMixin
 from . import utilities as utils
 
 class CheckBox(
@@ -10,6 +10,7 @@ class CheckBox(
     EnableMixin, 
     FocusMixin, 
     DisplayMixin, 
+    TextMixin,
     ReprMixin):
 
     def __init__(self, master, text, command=None, grid=None, align=None, args=None):